# api-doc.json
{
    "schemes": [],
    "swagger": "2.0",
    "info": {
...
      "title": "Mobile Risk API (MRA)",
      "contact": {
        "name": "esupport@lookout.com",
        "url": "https://esupport.lookout.com",
        "email": "esupport@lookout.com"
      },
      "version": "2.0"
    },
    "host": "api.lookout.com",
    "basePath": "/",
    "paths": {
...
      "/mra/api/v2/devices": {
        "get": {
          "description": "Get devices, filtered by specified filters\nSome filters like security_status, protection_status, state also take ! as a not equal prefix, e.g. `security_status=!SECURE` will find all devices with security_status != SECURE\n\nDevices are returned in sorted order, ordered by oid descending. To paginate or scroll through all devices, start with no oid in the request\nand a limit of under 1000, parse the response and then on subsequent requests, set the oid to be the last oid in the previous response and the same limit. Repeat\nuntil no more devices are returned in the response. For example, using a pagination limit of 2 devices a time (normally you'd use up to 1000 as a limit):\n\n```\ncurl -X 'GET' 'https://api.lookout.com/mra/api/v2/devices?limit=2' -H 'accept: application/json' -H \"Authorization: Bearer $ACCESS_TOKEN\"\n\n{\"count\":16,\"devices\":[{\"guid\":\"d56816b8-1d18-40e3-9500-42bc3378111f\",\"oid\":\"0190bc68-3917-76b3-98ab-991845b2d209\", ... }, {\"guid\":\"b19f66f7-520a-404a-9c4b-7bfc16ec5e8c\",\"oid\":\"0190bc67-3bf3-7417-988b-2f459393fa4b\", ... }]}\n```\nUsing the last oid in the previous response, continue to request additional data until the returned count is 0. For example:\n```\nurl -X 'GET' 'https://api.lookout.com/mra/api/v2/devices?oid=0190bc68-3917-76b3-98ab-991845b2d209&limit=2' -H 'accept: application/json' -H \"Authorization: Bearer $ACCESS_TOKEN\"\n\n{\"count\":14,\"devices\":[{\"guid\":\"6c6f3b14-84b9-4a4f-9ea3-071a9b06ac50\",\"oid\":\"0190bc69-00db-7b6f-9662-11809f1bf7b7\", ... }, {\"guid\":\"b2630992-b03a-45d9-955e-df323edc9b3b\",\"oid\":\"0190bc68-3917-76b3-98ab-991845b2d209\", ... }]}\n```\n\nThis is a rate-limited query, and may return 429 Too Many Requests if the service is under load. Wait to try again\nif the response status code is 429.",
          "produces": [
            "application/json"
          ],
          "tags": [
            "Devices"
          ],
          "summary": "Get devices by filters",
          "parameters": [
            {
              "type": "string",
              "default": "Bearer AccessToken",
              "description": "Bearer AccessToken",
              "name": "Authorization",
              "in": "header",
              "required": true
            },
            {
              "type": "string",
              "name": "aspl",
              "in": "query"
            },
            {
              "enum": [3, 6, 9, 12, 18, 24],
              "type": "integer",
              "name": "aspl_version_date",
              "in": "query"
            },
            {
              "type": "string",
              "name": "cve",
              "in": "query"
            },
            {
              "type": "boolean",
              "name": "device_admin_settings_activated",
              "in": "query"
            },
            {
              "type": "string",
              "description": "email, allows partial match",
              "name": "email",
              "in": "query"
            },
            {
              "type": "string",
              "format": "uuid",
              "description": "Device Group GUID",
              "name": "group",
              "in": "query"
            },
            {
              "type": "boolean",
              "name": "is_mdm",
              "in": "query"
            },
            {
              "maximum": 1000,
              "minimum": 0,
              "type": "integer",
              "description": "use oid and limit for pagination. Default: 100, Max: 1000",
              "name": "limit",
              "in": "query"
            },
            {
              "type": "string",
              "format": "uuid",
              "description": "Order by id. Used for pagination, returns devices with oid strictly \u003C the specified oid. The oid format uses a UUID v7 which is a time sortable version of a UUID",
              "name": "oid",
              "in": "query"
            },
            {
              "enum": [3, 6, 9, 12, 18, 24],
              "type": "integer",
              "name": "os_version_date",
              "in": "query"
            },
            {
              "type": "string",
              "name": "osversion",
              "in": "query"
            },
            {
              "type": "boolean",
              "name": "pcp_enabled",
              "in": "query"
            },
            {
              "type": "boolean",
              "name": "pcp_paused",
              "in": "query"
            },
            {
              "enum": [
                "ANDROID",
                "IOS"
              ],
              "type": "string",
              "name": "platform",
              "in": "query"
            },
            {
              "enum": [
                "WORK",
                "!WORK",
                "PERSONAL",
                "!PERSONAL"
              ],
              "type": "string",
              "name": "profile_type",
              "in": "query"
            },
            {
              "enum": [
                "PROTECTED",
                "!PROTECTED",
                "DISCONNECTED",
                "!DISCONNECTED"
              ],
              "type": "string",
              "name": "protection_status",
              "in": "query"
            },
            {
              "enum": [
                "THREATS_HIGH",
                "!THREATS_HIGH",
                "THREATS_MEDIUM",
                "!THREATS_MEDIUM",
                "THREATS_LOW",
                "!THREATS_LOW",
                "SECURE",
                "!SECURE"
              ],
              "type": "string",
              "name": "security_status",
              "in": "query"
            },
            {
              "enum": [
                "ACTIVATED",
                "!ACTIVATED",
                "DEACTIVATED",
                "!DEACTIVATED",
                "PENDING",
                "!PENDING"
              ],
              "type": "string",
              "name": "state",
              "in": "query"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/v2.GetDevicesResponse"
              }
            },
            "400": {
              "description": "Bad Request",
              "schema": {
                "$ref": "#/definitions/models.ErrorResponse"
              }
            },
            "401": {
              "description": "Unauthorized access - invalid or expired token, re-attempt after getting a new access token from the oauth2/token endpoint",
              "schema": {
                "$ref": "#/definitions/models.ErrorResponse"
              }
            },
            "429": {
              "description": "Too many requests, please wait a moment and try again.",
              "schema": {
                "$ref": "#/definitions/models.ErrorResponse"
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/models.ErrorResponse"
              }
            }
          }
        }
      },
...
    },
    "definitions": {
...
      "models.Client": {
        "type": "object",
        "properties": {
          "lookout_sdk_version": {
            "type": "string"
          },
          "ota_version": {
            "type": "string"
          },
          "package_name": {
            "type": "string"
          },
          "package_version": {
            "type": "string"
          }
        }
      },
...
      "models.Device": {
        "type": "object",
        "properties": {
          "activated_at": {
            "type": "string"
          },
          "activation_status": {
            "description": "oneof=ACTIVATED DEACTIVATED PENDING",
            "type": "string"
          },
          "checkin_time": {
            "type": "string"
          },
          "client": {
            "$ref": "#/definitions/models.Client"
          },
          "customer_device_id": {
            "type": "string"
          },
          "deactivated_at": {
            "type": "string"
          },
          "details": {
            "$ref": "#/definitions/models.DeviceDetails"
          },
          "device_group_guid": {
            "type": "string"
          },
          "device_permissions": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/models.DevicePermission"
            }
          },
          "device_settings": {
            "$ref": "#/definitions/models.DeviceSettings"
          },
          "device_vulns": {
            "$ref": "#/definitions/models.DeviceVulnerabilities"
          },
          "disconnected_at": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "enterprise_guid": {
            "type": "string"
          },
          "guid": {
            "type": "string"
          },
          "hardware": {
            "$ref": "#/definitions/models.Hardware"
          },
          "locale": {
            "type": "string"
          },
          "mdm_type": {
            "type": "string"
          },
          "oid": {
            "type": "string"
          },
          "parent_device_guid": {
            "type": "string"
          },
          "parent_status": {
            "$ref": "#/definitions/models.ParentStatus"
          },
          "platform": {
            "description": "oneof=ANDROID IOS",
            "type": "string"
          },
          "profile_type": {
            "description": "oneof=WORK PERSONAL if present",
            "type": "string"
          },
          "protection_status": {
            "description": "oneof=PROTECTED DISCONNECTED",
            "type": "string"
          },
          "risky_config": {
            "$ref": "#/definitions/models.RiskyConfig"
          },
          "security_status": {
            "description": "oneof=SECURE THREATS_LOW THREATS_MEDIUM THREATS_HIGH",
            "type": "string"
          },
          "software": {
            "$ref": "#/definitions/models.Software"
          },
          "updated_time": {
            "type": "string"
          }
        }
      },
      "models.DeviceAdminSettings": {
        "type": "object",
        "properties": {
          "activated": {
            "type": "boolean"
          }
        }
      },
...
      "models.DeviceDetails": {
        "type": "object",
        "properties": {
          "external_id": {
            "type": "string"
          },
          "mdm_connector_id": {
            "type": "integer"
          },
          "mdm_connector_uuid": {
            "type": "string"
          }
        }
      },
      "models.DevicePermission": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "value": {
            "type": "boolean"
          }
        }
      },
      "models.DeviceSettings": {
        "type": "object",
        "properties": {
          "device_admin_settings": {
            "$ref": "#/definitions/models.DeviceAdminSettings"
          },
          "pcp_settings": {
            "$ref": "#/definitions/models.PcpSettings"
          }
        }
      },
      "models.DeviceVulnerabilities": {
        "type": "object",
        "properties": {
          "vulnerabilities": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/models.VulnerabilityInfo"
            }
          }
        }
      },
      "models.ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "request-id": {
            "type": "string"
          }
        }
      },
...
      "models.Hardware": {
        "type": "object",
        "properties": {
          "manufacturer": {
            "type": "string"
          },
          "model": {
            "type": "string"
          }
        }
      },
...
      "models.ParentStatus": {
        "type": "object",
        "properties": {
          "activation_status": {
            "description": "oneof=ACTIVATED DEACTIVATED PENDING",
            "type": "string"
          },
          "parent_device_guid": {
            "type": "string"
          },
          "protection_status": {
            "description": "oneof=PROTECTED DISCONNECTED",
            "type": "string"
          },
          "security_status": {
            "description": "oneof=SECURE THREATS_LOW THREATS_MEDIUM THREATS_HIGH",
            "type": "string"
          }
        }
      },
...
      "models.RiskyConfig": {
        "type": "object",
        "properties": {
          "developer_mode": {
            "type": "string"
          },
          "developer_mode_v2": {
            "type": "string"
          },
          "encryption": {
            "type": "string"
          },
          "install_non_market_apps": {
            "type": "string"
          },
          "lock_screen": {
            "type": "string"
          },
          "packages_with_admin": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/models.PackagesWithAdmin"
            }
          },
          "risky_network_config": {
            "$ref": "#/definitions/models.RiskyNetworkConfig"
          },
          "trusted_signing_identities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "usb_debugging": {
            "type": "string"
          }
        }
      },
...
      "v2.GetDevicesResponse": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer"
          },
          "devices": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/models.Device"
            }
          }
        }
      },
...
    },
...
  }